+++ src/japverbconj/constants/irregular_verb_forms.py
@@ -53,6 +53,9 @@ SURU_CONJUGATION[BaseForm.TE][Formality.PLAIN][Tense.NONPAST][
 SURU_CONJUGATION[BaseForm.TE][Formality.POLITE][Tense.NONPAST][
     Polarity.POSITIVE
 ] = "しまして"
+SURU_CONJUGATION[BaseForm.TE][Formality.POLITE][Tense.NONPAST][
+    Polarity.NEGATIVE
+] = "しませんでして"
 
 SURU_CONJUGATION[BaseForm.TARI][Formality.PLAIN][Tense.NONPAST][
     Polarity.POSITIVE
@@ -63,6 +66,9 @@ SURU_CONJUGATION[BaseForm.TARI][Formality.PLAIN][Tense.NONPAST][
 SURU_CONJUGATION[BaseForm.TARI][Formality.POLITE][Tense.NONPAST][
     Polarity.POSITIVE
 ] = "しましたり"
+SURU_CONJUGATION[BaseForm.TARI][Formality.POLITE][Tense.NONPAST][
+    Polarity.NEGATIVE
+] = "しませんでしたり"
 
 SURU_CONJUGATION[BaseForm.CONDITIONAL][Formality.PLAIN][Tense.NONPAST][
     Polarity.POSITIVE
@@ -212,6 +218,9 @@ KURU_CONJUGATION[BaseForm.TE][Formality.PLAIN][Tense.NONPAST][
 KURU_CONJUGATION[BaseForm.TE][Formality.POLITE][Tense.NONPAST][
     Polarity.POSITIVE
 ] = "きまして"
+KURU_CONJUGATION[BaseForm.TE][Formality.POLITE][Tense.NONPAST][
+    Polarity.NEGATIVE
+] = "きませんでして"
 
 KURU_CONJUGATION[BaseForm.TARI][Formality.PLAIN][Tense.NONPAST][
     Polarity.POSITIVE
@@ -222,6 +231,9 @@ KURU_CONJUGATION[BaseForm.TARI][Formality.PLAIN][Tense.NONPAST][
 KURU_CONJUGATION[BaseForm.TARI][Formality.POLITE][Tense.NONPAST][
     Polarity.POSITIVE
 ] = "きましたり"
+KURU_CONJUGATION[BaseForm.TARI][Formality.POLITE][Tense.NONPAST][
+    Polarity.NEGATIVE
+] = "きませんでしたり"
 
 KURU_CONJUGATION[BaseForm.CONDITIONAL][Formality.PLAIN][Tense.NONPAST][
     Polarity.POSITIVE